Add summary metrics to saved json file (#46)
At the end of the run GuideLLM will print summary metrics that are
computed from the raw results, but these are not currently saved
anywhere.

This PR adds these metrics as serializable properties of the
`TextGenerationBenchmark` class. Most of the metrics are already
declared as properties, but are not serializable. Adding the
`@computed_field` decorator is enough in these cases. Other properties
were added to complete the list.