Add follow_redirects parameter to OpenAIHTTPBackend (#148)
This PR adds the ability to set `follow_redirects` on the httpx client
used to make requests to the OpenAI backend, allowing it to follow the
redirect URL in 3xx responses. Following a discussion with @markurtz,
this value defaults to true, but may be overwritten through environment
variables or the `OpenAIHTTPBackend` initializer.