fix links check fail, remove speculators from docs (#172)
We are getting link-check fails on the github rich display links. Not
sure what the desired solution is but this gets the check passing by
ignoring links like that.

Also we have what look to be some pasted doc remnants from the
speculators repo so I changed everything saying speculators to guidellm