Fix "Event loop Closed" error (#188)
By default httpx will keep the last few connections open in a pool for
re-use. Since we create the httpx connection inside an asyncio task, it
is bound to the lifetime of the task rather than the httpx client. Thus
when the client attempts to reuse the connection in a new task it fails.
There may be performance implications for this change, but since the
default is only to re-use the last 5 connections, its likely minor.
([ref](https://github.com/encode/httpx/discussions/2959))

Fixes: #147

---------

Co-authored-by: Mark Kurtz <mark.j.kurtz@gmail.com>