Option to re-display a benchmark file (#185)
closes #175 

This adds a command to re-display a prior benchmarks file in the CLI.

Before marking this as ready for review, we need to decide what command
format we want to use. During the call with Mark we discussed this being
an option within the benchmark command.
Also, let me know if the stripped down results file is a good one to
use. I manually removed data from a large results file.

---------

Co-authored-by: Samuel Monson <smonson@redhat.com>
Co-authored-by: Mark Kurtz <mark.j.kurtz@gmail.com>